feat:添加JPA增删改查基本操作

+++ src/main/java/com/haley/spring0to1/dao/DemoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface DemoRepository extends JpaRepository<Demo, Long> {
+public interface DemoRepository extends JpaRepository<Demo, Integer> {
 
 }
